this version has both a simplex method and convolution method working to find the best local offset for a given lattice node.
The simplex method works well, is stable and yields good results.
The convolution method also works, but the discretization of the 
possible deformed vector yields results that are not always stable.

The conv technique is only 2 times faster than the simplex method, 
and therefore does not warrent further testing....

I will now modify the simplex method:
   - use nearest neighbour interpolation only for the moving lattice
   - use tri-cubic for the non-moving lattice
   - swap the moving and stable grids.  now the source grid moves,
     the target (deformed) grid should move to increase the effective
     resolution of the deformation vector (over the linear interpolant
     limit).