This version of get_deformation_vector_for_node with the 3x3x3 weights of 1.0, 1.0, 1.0, 1.0 for all nodes used in the quadratic fit and works well with return_3D_disp_from_quad_fit() in ver 1.2 of quad_max_fit.c
This version of the quadratic fit seems to work reasonably well in 3D
and compares favorably to the simplex optimization for hoge/jacob fitting
at 16mm.

The use of quadratic fitting is twice as fast as simplex (6.4 vs 15.2 min)
for the 16mm fit (again with {hoge,jacob}_16_dxyz.mnc and -step 8 8 8.