new functional version of mincblur with new and improved default behavior. By default, only the blurred volume is created. If you want the gradient magnitude volumes, you have to ask for it (-gradient).
The temporary files (corresponding to the REAL data, and partial derivatives)
are removed by mincblur, so now it runs cleanly.  Unfortunately, these files
are _not_ deleted if mincblur fails, or is stopped.  I will have to use
unlink for this, but its a bit to much to do right now, since I would have
to change the way files are dealt with in gradient_volume.c, blur_volume.c
and gradmag_volume.c.

Also, it is possible to keep the partial derivative volumes (-partial).

this version is in mni_reg-0.1j