added comments for local fitting process.
changed the objective function used in the quadratic fitting routines to
be the same as that used for simplex, both should (and now do use)
xcorr_fitting_function() so that both the similarity of the data and the
cost of the deformation are taken into account.

started to add code for non-isotropic smoothing, but it is not completely
working in this version.