removed jiggle in do_nonlinear.c - The jiggle was causing multiple local minima in the local objective function.  This in turn was the MAJOR  CAUSE of errors in the estimation of the deformation field when using simplex optimization (and in fact was caused error for quad fit as well).
----
go_get_samples_in_source called with -1 for inter_type (before
the value for inter_type was ignored).

----
completed non-isotropic smoothing in return_locally_smoothed_def()
that is based on the eigen vectors of the hessian matrix (that is
derived from the differential geometry of the shape of the local
objective function).  The estimated deformation vectors are modulated
by a confidence function (see next para) such that the vectors are
smoothed less in the direction where the obj function has greater
curvature.

----
confidence_function has been implemented to give a confidence of
     -> 1.0 to eigen values >= the mean largest eigen_val (eval'd over
        the previous iteration)
     -> 0.0 to eigen values <= the mean smallest eigen_val
     -> [0.0 .. 0.5] to eigen_vals between smallest and middle mean
        eigen val
     -> [0.5 .. 1.0] to eigen_vals between middle and largest mean
        eigen val
This confidence_function is used to weight the amount of the additional
deformation vector added to the current def in return_locally_smoothed_def().