used inter_type to specify degress_continuity in the call to evaluate_volume_in_world() in procedure go_get_samples_in_source().
----
added support for tri-linear interpolation in
go_get_samples_with_offset() since the local objective function is not
smooth (but is actually step-wise) with NN interpolation.  This is
specified by an extra boolean parameter to go_get_samples_with_offset
that should be TRUE to use NN interpolation.  This is controled from
the command line by the -nearest_neighbour / -trilinear options.

NOTE: NN is faster, but the local obj function will have steps (ie it
      will have plateaus and discontinuous changes).  Also, there
      seems to be a `drapery' effect, such that the function is _not_
      monotonically increasing with mis-registration -> thus yielding
      local minima.
      -> trilin is the default, and should be used unless you really
         know what you are doing!

----

removed dx += 0.5; dy += 0.5; dz += 0.5.; in
go_get_samples_with_offset().  this was originally supposed to give a
`rounding' effect for NN interpolation, but is no longer needed with
tri-linear interpolation.  IN FACT: keeping it caused a certain bias
so that a deformation was returned _even_ when an object was matched
to itself; after removal- there are only small sporadic deviations
from a null deformation field.